kubernetes-backend: Moves backend specific types back from common
