+++ plugins/kubernetes-backend/src/types/types.ts
@@ -14,18 +14,116 @@
  * limitations under the License.
  */
 
-export type {
-  ClusterDetails,
-  CustomResource,
+import type {
   FetchResponse,
-  FetchResponseWrapper,
-  KubernetesClustersSupplier,
+  KubernetesFetchError,
+} from '@backstage/plugin-kubernetes-common';
+
+export type {
+  FetchResponse,
   KubernetesErrorTypes,
   KubernetesFetchError,
-  KubernetesFetcher,
-  KubernetesObjectTypes,
   KubernetesRequestBody,
-  KubernetesServiceLocator,
-  ObjectFetchParams,
-  ServiceLocatorMethod,
 } from '@backstage/plugin-kubernetes-common';
+
+export type ClusterLocatorMethod =
+  | ConfigClusterLocatorMethod
+  | GKEClusterLocatorMethod;
+
+export interface ConfigClusterLocatorMethod {
+  /**
+   * @visibility frontend
+   */
+  type: 'config';
+  clusters: {
+    /**
+     * @visibility frontend
+     */
+    url: string;
+    /**
+     * @visibility frontend
+     */
+    name: string;
+    /**
+     * @visibility secret
+     */
+    serviceAccountToken: string | undefined;
+    /**
+     * @visibility frontend
+     */
+    authProvider: 'aws' | 'google' | 'serviceAccount';
+  }[];
+}
+
+export interface GKEClusterLocatorMethod {
+  /**
+   * @visibility frontend
+   */
+  type: 'gke';
+  /**
+   * @visibility frontend
+   */
+  projectId: string;
+  /**
+   * @visibility frontend
+   */
+  region?: string;
+}
+
+export interface CustomResource {
+  group: string;
+  apiVersion: string;
+  plural: string;
+}
+
+export interface ObjectFetchParams {
+  serviceId: string;
+  clusterDetails: ClusterDetails;
+  objectTypesToFetch: Set<KubernetesObjectTypes>;
+  labelSelector: string;
+  customResources: CustomResource[];
+}
+
+// Fetches information from a kubernetes cluster using the cluster details object
+// to target a specific cluster
+export interface KubernetesFetcher {
+  fetchObjectsForService(
+    params: ObjectFetchParams,
+  ): Promise<FetchResponseWrapper>;
+}
+
+export interface FetchResponseWrapper {
+  errors: KubernetesFetchError[];
+  responses: FetchResponse[];
+}
+
+// TODO fairly sure there's a easier way to do this
+
+export type KubernetesObjectTypes =
+  | 'pods'
+  | 'services'
+  | 'configmaps'
+  | 'deployments'
+  | 'replicasets'
+  | 'horizontalpodautoscalers'
+  | 'ingresses'
+  | 'customresources';
+
+// Used to load cluster details from different sources
+export interface KubernetesClustersSupplier {
+  getClusters(): Promise<ClusterDetails[]>;
+}
+
+// Used to locate which cluster(s) a service is running on
+export interface KubernetesServiceLocator {
+  getClustersByServiceId(serviceId: string): Promise<ClusterDetails[]>;
+}
+
+export type ServiceLocatorMethod = 'multiTenant' | 'http'; // TODO implement http
+
+export interface ClusterDetails {
+  name: string;
+  url: string;
+  authProvider: string;
+  serviceAccountToken?: string | undefined;
+}
